first pass at wq/store.js overhaul (fixes #20)
 - use localforage to handle actual storage
 - use promises (fixes #5)
 - require async usage (fixes #17)
 - split getList() into separate wq/model.js module
 - split outbox stuff into separate wq/outbox.js module
 - simplify init (see #28)
